Restyle headings format only on the changed line instead of whole text
Start doing partial (not fully incremental) formatting instead of
restyling the whole document. The advantage of this method should be
obvious: The time taken for restyling will become proportional to the
size of current line being edited and not the whole document. As a
result, it will result a major speedup for large notes, specially those
which has many formattings.

The ultimate solution is to bring in a parser, perhaps an incremental
one, but as it requires time and effort, so this is good to have. Plus,
partial restyling is something that would be needed in the case of a
non-incremental parser.