Add option for wrapping of long note titles.
This introduces a new setting long-note-names-handling with valid values of -1 (placeholder), 0 (ellipsize), 1 (wrap) and 2 (expand the note list column). When the application starts, if long-note-names-handling is set to -1, then the existing long-note-names boolean setting is imported, setting long-note-names-handling to either 0 or 2. The switch in the preferences dialog is replaced with a combo box to select from the three options.

Closes #285